pyproject.toml
[project]
name = "llm-eval"
version = "0.1.0"
description = "LLM Response Evaluation Framework — cross-language comparison"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40",
    "openai>=1.50",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "typer>=0.12",
    "rich>=13.0",
    "pyyaml>=6.0",
    "sentence-transformers>=3.0",
    "numpy>=1.26",
    "matplotlib>=3.8",
]

[project.scripts]
llm-eval = "llm_eval.cli:app"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_eval"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
